// sdk/ai/azure-ai-client/src/models/models.ts
export interface ConnectionsListResponse {
  /** A list of connection list secrets */
  value: ConnectionsListSecretsResponse[];
}

/** Response from the listSecrets operation */
export interface ConnectionsListSecretsResponse {
  /** The name of the resource */
  name: string;
  /** The properties of the resource */
  properties: ConnectionPropertiesUnion;
}

/** Connetion properties */
export interface ConnectionProperties {
  /** the discriminator possible values: ApiKey, AAD, SAS */
  authType: AuthenticationType;
}

/** Connection properties for connections with API key authentication */
export interface ConnectionPropertiesApiKeyAuth extends ConnectionProperties {
  /** Authentication type of the connection target */
  authType: "ApiKey";
  /** Category of the connection */
  category: EndpointType;
  /** Credentials will only be present for authType=ApiKey */
  credentials: CredentialsApiKeyAuth;
  /** The connection URL to be used for this service */
  target: string;
}

/** The Type (or category) of the connection */
export type EndpointType = "AzureOpenAI" | "Serverless";

/** The credentials needed for API key authentication */
export interface CredentialsApiKeyAuth {
  /** The API key */
  key: string;
}

/** Connection properties for connections with AAD authentication (aka `Entra ID passthrough`) */
export interface ConnectionPropertiesAADAuth extends ConnectionProperties {
  /** Authentication type of the connection target */
  authType: "AAD";
  /** Category of the connection */
  category: EndpointType;
  /** The connection URL to be used for this service */
  target: string;
}

/** Connection properties for connections with SAS authentication */
export interface ConnectionPropertiesSASAuth extends ConnectionProperties {
  /** Authentication type of the connection target */
  authType: "SAS";
  /** Category of the connection */
  category: EndpointType;
  /** Credentials will only be present for authType=ApiKey */
  credentials: CredentialsSASAuth;
  /** The connection URL to be used for this service */
  target: string;
}
// ...
export type ConnectionPropertiesUnion =
  | ConnectionPropertiesApiKeyAuth
  | ConnectionPropertiesAADAuth
  | ConnectionPropertiesSASAuth
  | ConnectionProperties;